Finally working Catalog still no Filter by category but the products are showing also moved users.js to user.js to match database
