+++ server/api.js
@@ -7,7 +7,7 @@ const api = module.exports = require('express').Router()
 api
     .get('/heartbeat', (req, res) => res.send({ok: true}))
     .use('/auth', require('./auth'))
-    .use('/users', require('./users'))
+    .use('/user', require('./user'))
     .use('/catalog', require('./catalog'))
 
 // No routes matched? 404.
